Bug. Download everything
